Move header_stream into decrypt_stream
Everything left in header_stream is strictly for decryption
